[CI] Simplify Python test execution
Run the regular Python suite from its canonical test root while preserving load-group scheduling for fixture ordering. Keep test-only pytest hooks local, move NNAPI tests under the nightly tree, and remove inert launchers and stale platform exclusions.
