+++ ci/jenkins/templates/gpu_jenkinsfile.groovy.j2
@@ -53,10 +53,6 @@
     script: "${docker_run} ${ci_gpu} ./tests/scripts/task_python_unittest_gpuonly.sh",
     label: 'Run Python GPU unit tests',
   )
-  sh (
-    script: "${docker_run} ${ci_gpu} ./tests/scripts/task_python_integration_gpuonly.sh",
-    label: 'Run Python GPU integration tests',
-  )
 {% endcall %}
