+++ tests/python/disco/test_custom_allreduce.py
@@ -26,7 +26,6 @@ from tvm_ffi import Shape
 import tvm
 import tvm.testing
 from tvm.runtime import DataType, disco
-from tvm.runtime.disco import Session
 
 
 class AllReduceStrategyType(enum.IntEnum):
@@ -56,7 +55,7 @@ _ccl = [ccl for ccl in _compiled_ccl() if ccl == "nccl"]
 @pytest.mark.parametrize("strategy", _strategies)
 def test_allreduce(shape, ccl, strategy):
     devices = [0, 1]
-    sess: Session = disco.ProcessSession(num_workers=len(devices))
+    sess: disco.Session = disco.ProcessSession(num_workers=len(devices))
     sess.init_ccl(ccl, *devices)
 
     num_elements = reduce(lambda x, y: x * y, shape)
