+++ tests/python/tirx/conftest.py
@@ -25,6 +25,8 @@ real sm_100a device so it skips cleanly where the hardware is absent and runs
 in full where it is present.
 """
 
+from pathlib import Path
+
 import pytest
 
 from tvm.testing import env
@@ -33,8 +35,10 @@ from tvm.testing import env
 def pytest_collection_modifyitems(config, items):
     if env.has_cuda_compute(10):
         return
+    suite_root = Path(__file__).parent
     skip = pytest.mark.skip(
         reason="tirx suite requires a CUDA compute capability 10.0 (sm_100a) device"
     )
     for item in items:
-        item.add_marker(skip)
+        if item.path.is_relative_to(suite_root):
+            item.add_marker(skip)
